Fix type overflow issue in unread count handling
Changed UnreadCount type from int to int64 throughout the codebase to prevent
potential overflow on 32-bit systems or with very large unread counts.

Changes:
- Backend: DirectRoomResponse.UnreadCount changed from int to int64
- Backend: Removed unsafe int() type conversion in buildDirectRoomResponse
- CLI: DirectRoom.UnreadCount changed from int to int64

This ensures type consistency as GetUnreadCountsForUser returns map[uint]int64,
and Room.UnreadCount is already int64. All unread counts now use int64 without
any unsafe conversions.

+++ chat-backend-go/handlers/room_handlers.go
@@ -37,7 +37,7 @@ type DirectRoomResponse struct {
 	DisplayName string           `json:"display_name"`
 	OtherUser   OtherUserInfo    `json:"other_user"`
 	LastMessage *LastMessageInfo `json:"last_message,omitempty"`
-	UnreadCount int              `json:"unread_count"`
+	UnreadCount int64            `json:"unread_count"`
 	CreatedAt   time.Time        `json:"created_at"`
 }
 
@@ -97,7 +97,7 @@ func buildDirectRoomResponse(room models.Room, currentUserID uint, lastMsgMap ma
 			IsOnline:     otherUser.IsOnline,
 			LastActiveAt: otherUser.LastActiveAt,
 		},
-		UnreadCount: int(unreadCounts[room.ID]), // Get actual unread count from map
+		UnreadCount: unreadCounts[room.ID], // Get actual unread count from map (int64)
 		CreatedAt:   room.CreatedAt,
 	}
